+++ test/fast_count_test.rb
@@ -6,6 +6,7 @@ class FastCountTest < Minitest::Test
   def setup
     FastCount.install
     @connection = ActiveRecord::Base.connection
+    @connection.schema_cache.clear!
   end
 
   def teardown
@@ -84,6 +85,51 @@ class FastCountTest < Minitest::Test
 
   def test_estimated_count
     assert_kind_of Integer, User.where(admin: true).estimated_count
+  end
+
+  def test_fast_distinct_count_raises_if_missing_index
+    skip unless postgresql? || mysql?
+
+    error = assert_raises(RuntimeError) do
+      User.fast_distinct_count(column: :company_id)
+    end
+    assert_match("Index starting with 'company_id' must exist on 'users' table", error.message)
+  end
+
+  def test_fast_distinct_count_raises_if_incorrect_index_exists
+    skip unless postgresql? || mysql?
+
+    begin
+      @connection.add_index(:users, [:name, :company_id]) # 'company_id' is not a prefix
+      error = assert_raises(RuntimeError) do
+        User.fast_distinct_count(column: :company_id)
+      end
+      assert_match("Index starting with 'company_id' must exist on 'users' table", error.message)
+    ensure
+      @connection.remove_index(:users, [:name, :company_id])
+    end
+  end
+
+  def test_fast_distinct_count
+    @connection.add_index(:users, :company_id)
+    @connection.execute(<<~SQL)
+      INSERT INTO users (company_id) VALUES (1), (2), (3), (2), (2), (3), (4)
+    SQL
+
+    assert_equal 4, User.fast_distinct_count(column: :company_id)
+  ensure
+    @connection.remove_index(:users, :company_id)
+  end
+
+  def test_fast_distinct_count_counts_nulls
+    @connection.add_index(:users, :company_id)
+    @connection.execute(<<~SQL)
+      INSERT INTO users (company_id) VALUES (1), (2), (3), (2), (null), (null), (4)
+    SQL
+
+    assert_equal 5, User.fast_distinct_count(column: :company_id)
+  ensure
+    @connection.remove_index(:users, :company_id)
   end
 
   private